net: shell: Add mem command
Move "mem" command to a separate file.
